feat: udistrital/sga_documentacion#167 Se pasa la lógica a legalizacióm.service
